Small code refactor
* Rework acl tests to use memory storage instead of an actual zdb
* Check if previous test file exists before trying to remove it in sql
meta tests
* Add an expect to the result returned when clearing a namespace in zdb,
and move the function in an impl which is only compiled in test runs
* Rework bcdb service to take a generic object store

+++ src/meta/sqlite.rs
@@ -210,7 +210,9 @@ mod tests {
     #[tokio::test]
     async fn schema() {
         let db = "/tmp/testing.sqlite3";
-        std::fs::remove_file(db).expect("failed to clean up file");
+        if std::path::Path::new(db).exists() {
+            std::fs::remove_file(db).expect("failed to clean up file");
+        }
 
         let constr = format!("sqlite://{}", db);
         let c = SqlitePool::new(&constr).await.expect("failed to connect");
@@ -243,11 +245,5 @@ mod tests {
                 println!("{}: {}", t.key, t.value);
             }
         }
-        // use tokio::stream::StreamExt;
-        // tokio::pin!(cur);
-        // //tokio::spawn(async move {});
-        // while let Some(row) = cur.next().await {
-
-        // }
     }
 }
